f1_strategy_system/requirements.txt
# ── Data Collection ──────────────────────────────
fastf1==3.3.9
requests==2.31.0
beautifulsoup4==4.12.3
aiohttp==3.9.3

# ── Data Processing ───────────────────────────────
pandas==2.2.0
numpy==1.26.4
pyarrow==15.0.0
pyspark==3.5.1

# ── Stream Processing ─────────────────────────────
kafka-python==2.0.2
apache-flink==1.18.0; platform_system=="Linux"

# ── Causal Inference (FCSG) ───────────────────────
causal-learn==0.1.3.8
dowhy==0.11.1
econml==0.15.0
networkx==3.2.1

# ── Federated Learning (FCSG) ─────────────────────
flwr==1.7.0

# ── Bayesian Optimization (ABPWO) ─────────────────
torch==2.2.0
botorch==0.10.0
gpytorch==1.11

# ── Reinforcement Learning (MARL) ─────────────────
ray[rllib]==2.9.3
gymnasium==0.28.1

# ── NLP / LLM (CSIE) ──────────────────────────────
langchain==0.1.9
anthropic==0.18.1
openai==1.12.0

# ── API Serving ───────────────────────────────────
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1

# ── Experiment Tracking ───────────────────────────
mlflow==2.10.2

# ── Visualization & Notebooks ─────────────────────
jupyter==1.0.0
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.19.0

# ── Utilities ─────────────────────────────────────
python-dotenv==1.0.1
pyyaml==6.0.1
tqdm==4.66.2
loguru==0.7.2
